Add device manager and fix shining spring(?) bug.
A hash table is used to find registered devices. An ID number is
allocated to each device to be distinguished. And `stdin`, `stdout`,
`stderr` are defined.

I was hesitating to implement device manager because I hadn't yet
decided whether to use the notion of file system, how to represent a
device or any kind of data to a user. As doing homework things can be
changed later.

system call bug fix:

  r0 must be the same value to the stacked one because of hardware
  mechanism. The meantime of entering into interrupt service routine,
  nothing can change the value. But the problem is that it sometimes
  gets polluted. how? why? I haven't yet figured out the reason. So I
  just get r0 from stack. Below is the code in /mach/stm32/irq.c:

	  ldr	r0, [sp] --> this line added
	  cmp	r0, %0
	  it	ge
	  ...

memory manager bug fix:

  Initial kernel stack would have overlaped with very early allcocation
  requests such as usart fifo.

  Page array mem_map and bitmap region were not marked as allocated.

scheduler bug fix:

  When only a task runnable, it was replaced with null by scheduler.

What a beautiful day anyway.

+++ include/kernel/task.h
@@ -47,12 +47,12 @@ struct sched_entity {
 } __attribute__((packed));
 
 struct task_t {
-	int state;       	/* priority */
+	long state;       	/* priority */
 	unsigned long primask;
 
-	unsigned *stack; 	/* stack end, the lowest address */
-	unsigned *sp; 		/* stack pointer */
-	int stack_size;  	/* stack size */
+	unsigned long *stack; 	/* stack end, the lowest address */
+	unsigned long *sp;	/* stack pointer */
+	long stack_size;  	/* stack size */
 
 	void *addr;      	/* address */
